fix: refactors config loaders
