+++ src/aind_data_transfer/jobs/openephys_job.py
@@ -11,7 +11,9 @@ from pathlib import Path
 
 from aind_codeocean_api.codeocean import CodeOceanClient
 
-from aind_data_transfer.configuration_loader import EphysJobConfigurationLoader
+from aind_data_transfer.config_loader.ephys_configuration_loader import (
+    EphysJobConfigurationLoader,
+)
 from aind_data_transfer.readers import EphysReaders
 from aind_data_transfer.transformations.compressors import EphysCompressors
 from aind_data_transfer.transformations.metadata_creation import (
